feat(migrations): update categories and products tables
- Added 'price' and 'discount' fields to the categories table.
- Refactored products table to use 'id()' method for primary key and added 'code', 'sold', and 'expiration_date' fields.
- Changed 'product_id' in products_categories table to use 'unsignedBigInteger' for better compatibility with the products table.

+++ database/migrations/2025_03_20_002810_create_products_categories_table.php
@@ -11,7 +11,7 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::create('products_categories', function (Blueprint $table) {
-            $table->unsignedInteger('product_id')->nullable(false);
+            $table->unsignedBigInteger('product_id')->nullable(false);
             $table->foreign('product_id')
                 ->references('id')
                 ->on('products')
